[IMP] cetmix_tower_server
This is the first attempt to implement file push in flight plans.
This is done with a new type of action in flight plan lines.

Why it's abandoned: it's more handy to have file uploading at the command level.
Because it allows to upload files using commands not only flight plans.