[IMP] cetmix_tower_server Command duration
Before
------
Command duration was computed once the a command has finished.
So it was shown as '0' as long as a command was running.

Now
---
If a command is still duration is computed as a difference between command start time and current time.
For finished commands duration is show as a total consumed.