[FIX] cetmix_tower_server Command log access test
Access right and rules are checked during `search` and `read` operations.
When `.` dot notation is used ORM may use some of those method however this is not guaranteed because of intensive usage of cache.
This is why we are using `read` method explicitly to ensure access rights/rules are checked.