[IMP] cetmix_tower_server Server logs
Server Log is a special model that is designed to improve server log management UX.

Server logs can be fetched two different ways:
- Using a command
- From a file

Both options are configurable in the log settings.
A new tab "Server Logs" is added on the server form. From here user can configure and check the logs.
User can click on the log line to open it in a pop up form. Or click the "Open" button to open a log in a regular form view.

Importnat note:

Command flow is modified in order to be able to return command output instead of logging it in the `Command Log` record.