[IMP] cetmix_tower_server Variable values update
Before this commit
------------------

- 'Value' field was required
- 'Variable' field was not required

After this commit
-----------------

- 'Value' field is not required
- 'Variable' field is required

Why?
----
This is needed to allow overriding global variable values locally
with 'False' value.
Also there cannot be a variable value without a variable,
therefore 'Variable' must be a mandatory field.

Example
-------
Variable 'branch' has a global value 'main'.
There is a command that creates a new branch on;y if branch is
not defined (==False).
Before this commit the only way to achieve that was to assign some value
eg "no" and check for it: `if {{ branch }} == 'no':`
After this commit value field may remain empty so the same check
will look like: `if not {{ branch }}:`

Task: 4030