[IMP] cetmix_tower_server: Store related keys
Improve the way related keys are stored in commands.
Store only keys that either global or connected to command servers
or to partners connected to command servers.

Store related secrets in the file templates.

Modify cx.tower.key constraints to ensure that they
are working as expected.
SQL constraint was useless because both 'partner' and 'server' fields
allow NULL values.
So an ORM constraint was added to ensure that the key is unique.
Delete the constrain inherited from mixin after committing database schema.

Task: 4166