[FIX] cetmix_tower_git: Related records creation
When a record is created, some related records are created as well.
Some of them use rules that depend on `user_ids` and `manager_ids` fields.
However at the moment of record creation, these fields are not yet set.
So first we create the record without these related records, then we create
the related records to avoid access violations.

Task: 4369