[IMP][BREAKING] cx_tower_server: Refactor secrets
Before this commit
-------------------

Current implementation of secrets presumes that a secret can be linked
either to a server or a partner.
This requires to make an exclusion from the unique constraint for the
references to allow to have several secrets with the same reference.
This breaks the general paradigm of references being unique, raises
confusion and is error prone.

After this commit
-----------------

New dedicated model is added to store custom secret values for secrets.
At the same time backward compatibility is kept by converting "Secret Value"
field of a secret into a computed one.

Now there is no need in custom reference generation for secrets.
Also it's much simpler to understand and maintain secrets logic and
compute "Related Secrets" field in Commands, Files and File Templates.

Task: 4413