[IMP] cetmix_tower_server: variable value modifier
Allow to modify variable values using Python code.

Why?
----

In case we need to post-process the variable value, for example to
remove special characters or modify it based on some logic.
Eg. we have a variable which values are used for URLs, so we want to
remove all special characters from the value.

Task: 4474