[IMP] cetmix_tower_server: Custom variable values
UX
--

Add ability to pass custom variable values to commands.
This is needed to override values that are rendered from the server.

**Important**: user must have `write` access to the server in order to
be able to submit custom configuration values.
Otherwise, only values rendered from the server will be used.

UI
--

Add a new "Configuration Values" page to the command run wizard.
This page allows to edit variable values passed to the command.
Only users with `write` access to the server can edit values in the page.

Possible use cases
------------------

A command that updates modules:

- Before: list of modules is defined in the server configuration values.
- After: list of modules can be provided dynamically to the command.

A command that creates a backup of a database:

- Before: database name is defined in the server configuration values.
- After: database name can be provided dynamically to the command.

Task: 4524