[REF] cetmix_tower_server: Command and Flight Plan
Before this commit
------------------

When a user tried to run a command or a flight plan, that were incompatible
with the server and exception was raised.
This worked fine when a command or a flight plan were launched from the UI.
However this created issues when a command or a flight plan were launched
from code because a log record was not created and the only way to find
out what happened was to check the Odoo server logs.
It also created issues with the third party runners, such as queue_job,
causing them to fail without leaving any trace in the Tower logs.

After this commit
-----------------

1. Incompatible commands and flight plans are creating log records with
   the corresponding status.
2. New field is added to the Flight Plan Log model to allow storing
   a custom message.
3. Improve the flight plan 'copy' method to handle nested records.
4. Do some minor code cleanup.

Task: 4749