[REF] cetmix_tower_server Python command
Improve the way Python libraries are added using custom modules.
This is needed to be able to track which module is adding which library.
So we can later include those modules in YAML file dependencies when
exporting commands containing those libraries.

For example, when we use the 'boto3' library that is added by the
'cetmix_tower_aws' module in our command and later export that command,
we should ensure that the destination Cetmix Tower instance where this
command will be imported will have the 'cetmix_tower_aws' module installed
as well.

Task: 4832