[16.0][IMP]cetmix_tower_server: variable widget
This commit improves the developer experience when editing `code` fields in models using the template mixin.

Why this is needed
------------------
Previously, when adding variables to a code field (e.g. in a command), users had to manually:
- Search for the desired variable
- Copy its reference
- Insert it manually as `{{ some_variable }}`

New behavior
------------
Typing `{{` inside the code editor now triggers a dropdown with all available variables...

Task: 4827