[16.0][IMP]cetmix_tower_server: secret autocomplete in editor
This commit adds support for autocompleting secret references in the
"ace_tower" code editor. Previously, users had to manually find and
insert the reference to a secret in the format
#!cxtower.secret.secret_name!#.

With this update, typing the #! sequence inside the editor triggers a
dropdown list of available secrets. Selecting a secret from the list
inserts the full reference with the required trailing !# symbols.

This improves UX by speeding up the process and reducing the chance of
errors when referencing secrets in code or command templates.

Task: 4853