[IMP] cetmix_tower_server References in files
Before this commit references for files were generated based on
the file name. This was too generic and could lead to conflicts
when importing servers together with files.
Because there could be another file with the same and reference belonging
to another server.

After this commit references are generated based on the server reference
thus minimizing the risk of conflicts.

File views are modified:

- List view: removed the file content field, added the reference field
- Form view: reference field is not shown for new records. This is done
  to give priority to the auto generated reference.
  However, the reference can be edited later if needed.
- Search view: search by reference

Task: 4870