[FIX] cetmix_tower_yaml File upload
Before this fix when a user that was not an Odoo "Access Rights" group member
tried to upload a YAML file, the import wizard would fail with error:
"You are not allowed to access 'Models' (ir.model) records.".

This fix allows the import wizard to be used by any user, even if they are not
an Odoo "Access Rights" group member.

Task: 4899