[IMP] cetmix_tower_server: server log
Use text field instead of html field for log text.

**Why?**

- Quicker to render
- Easier to extend later in customizations

Add helper functions to get log text and format it. This also allows to
get log text from source without saving it to the database.
This can be useful for the cases when log should be displayed and refreshed
without saving it to the database.

Task 4909