[FIX] cetmix_tower_server Command exception
Handle exceptions while running SSH commands.
Before this commit and exception in the SSH command executed as a part
of a child flight plan lead to the situation where the child plan
remained stuck in the running state.
After this commit and exception in the SSH command will be handled
correctly and the child plan will be finished with an error.

Task: 5009