[IMP] cetmix_tower_server_queue: command finish
Before this commit the 'finish()' method was written to handle only one
command log at a time.
However in the parent class this method is designed to handele recordsets.
This lead to errors when this method was triggered for multiple command logs.
Eg when killing 'zombie' commands.

This commit fixes the issue by filtering out the command logs that are already
stopped and making it recordset-friendly.

Task: 5062