[IMP][BREAKING] cetmix_tower_yaml Views
Before this commit there was a dedicated 'YAML Export' page in some views.
Also there was a dedicated 'YAML Export' action in the list views.

However some models lack notebook so there was no page while the action
was still present.

This commit removes the dedicated 'YAML Export' page and action from the views.

Instead, the 'YAML Export' action is now present in both list and form views.
This ensures consistency and fixes the issue with  models lacking a notebook.

Task: 5114