[IMP] cetmix_tower_server: Key parsing
NB: this is actually not a bug but an expected behaviour.
However it turns out that it creates issues.

Before this commit:
-------------------

`if x == #!cxtower.secret.postgres_password!#`
where "Postgres Password" is not defined is parsed as

`if x == #!cxtower.secret.postgres_password!#`

This breaks Python code execution and leads to error.

After this commit:
-------------------
`if x == #!cxtower.secret.postgres_password!#`

where "Postgres Password" is not defined is parsed as

`if x == None`

This ensures that the code is executed correctly.

Task: 5134